refactor: refactor register page (#46)
